feat: added buying functionality

+++ src/main/java/com/example/noservendingmachine/dto/money/Banknote.java
@@ -1,7 +1,7 @@
 package com.example.noservendingmachine.dto.money;
 
 public abstract class Banknote extends Money {
-    protected Banknote(final int value) {
+    protected Banknote(final float value) {
         super(value);
     }
 }
