+++ src/main/java/com/example/noservendingmachine/dto/money/Coin.java
@@ -1,7 +1,7 @@
 package com.example.noservendingmachine.dto.money;
 
 public abstract class Coin extends Money {
-    protected Coin(final int value) {
+    protected Coin(final float value) {
         super(value);
     }
 }
